updated the jewelcli because I needed better optional argument support.
added the --neo4jconfig option as optional and if given uses the Neo4jGraph; if
it is not there the program defaults to the InMemoryGraph.

the --out option is still valid but optional now as well. if it is not provided no output is written (if InMemoryGraph is used than nothing happens.)

updated all the converters to allow this and support the Neo4jGraph.

neo4j supports named relationship (edge) types. This was supported via the Interaction property (and is still there). Neo4j now just uses that value as the relationshiptype. this has to be done at edge creation time and therefor I added a addEdge(..., relationshiptypename, relationshipvalue) methods. For the InMemoryGraph it just sets the Interaction Attribute.